requirements.txt
comet_ml
git+https://github.com/openai/CLIP.git
pandas
h5py
pytorch_transformers
decord